+++ src/html/stripe/src/charge.ts
@@ -1,8 +1,8 @@
 import { loadStripe } from '@stripe/stripe-js/pure'
 
-import { mount } from '../../mount.js'
 import type { Methods } from '../../../stripe/index.js'
 import type { HtmlConfig } from '../../../stripe/server/Charge.js'
+import { mount } from '../../mount.js'
 
 mount<typeof Methods.charge, HtmlConfig>(async (c) => {
   const request = c.challenge.request